feat(discord-notification): Update release notification message
The changes update the Discord release notification message to be more
engaging and informative. The previous message of "@everyone" is
replaced with a more descriptive and inviting message "Certified
fresh… until the next release". This will also ensure that we do not 
disturb everyone in the community for frequent releases and only on 
biweekly releases