feat: Add framework/library description field
The changes add a new `description` field to the `FrameworkModel` class in the `unoplat_code_confluence_commons` module. This field allows for a textual description of the framework or library to be stored, providing more context and information about the resource.

Additionally, the version of the `unoplat-code-confluence-commons` package is updated from `0.25.0` to `0.26.0` to reflect the new feature addition.