refactor(parser): Remove linting and processing code - fighting software complexity
we had added linting in a way that will improve the user codebase in a 
way that would make it simpler for us to write s expression quries to 
extract signals we want for agents.md. the downside was for our agents 
we had to give them search pattern instead of start line end line which 
becomes harder for them in terms of how to search finite element as they
will not know the end location. we are reversing this and will take upon
grammar changes required but will make it easy/reliable/faster for our 
Agents